No more memoizing valid_*_attributes, deprecated old practice.
The practice of passing a single argument to the builder block
and getting access to the overrides hash, thereby being able to
instantiate the instance directly on the class has been deprecated.

My reason for this is that it's unexpected to have different things
in the first argument position based on arity. Also, the merging
proxy is much more useful than the overrides hash, since it always
merges the attributes, and allows the declaration of protected
attributes, whereas you don't even always need to process overrides.